backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
openai==1.3.0
apscheduler==3.10.4
python-dotenv==1.0.0
yfinance==0.2.18
python-multipart==0.0.6